Rename sft_config to runner_config
