+++ examples/demo-rollup/src/main.rs
@@ -5,8 +5,8 @@ use const_rollup_config::{ROLLUP_NAMESPACE_RAW, SEQUENCER_DA_ADDRESS};
 use demo_stf::app::DefaultContext;
 use demo_stf::app::{DefaultPrivateKey, NativeAppRunner};
 use demo_stf::genesis_config::create_demo_genesis_config;
+use demo_stf::runner_config::from_toml_path;
 use demo_stf::runtime::GenesisConfig;
-use demo_stf::stf_config::from_toml_path;
 use jupiter::da_service::CelestiaService;
 use jupiter::types::NamespaceId;
 use jupiter::verifier::CelestiaVerifier;
