+++ examples/demo-stf/src/lib.rs
@@ -1,9 +1,9 @@
 pub mod app;
 #[cfg(feature = "native")]
 pub mod genesis_config;
-pub mod runtime;
 #[cfg(feature = "native")]
-pub mod stf_config;
+pub mod runner_config;
+pub mod runtime;
 #[cfg(test)]
 pub mod tests;
 pub mod tx_hooks_impl;
